refactor: extract HTTP server helpers
